feat: add build command
